NET-68: review fix
